Incorporate Firebase Analytics and update Firebase configurations
This commit incorporates Firebase Analytics into different components within the app for better event tracking. Furthermore, the Firebase project configurations for both development and production environments have been updated and the sign-in options are modified in the app module. Some changes in .firebaserc show an update in Firebase deployment configurations. An Angular Firebase link is added to the README file.